ACTUALLY fixed echo_debug.
As in, this time I actually tested the code instead of just assuming it worked like I tend to do sometimes -_- .

Also added some temp sprites, the coin thing I spent a little too much time on, but it was just some simple mathematics to get the shading. And the extra life is just my shitty - less than 30 second - drawing of a 1 up mushroom. All I need to add now is a temp sprite for keys. Which I could have done before committing but it's almost 11 PM so yeahhhhhh.