new inventory sprite
also fixed the inventory selector being able to get one more spaces than it should have.

Also the spike_wall_l.png file was changed, but it will be changed back, just wanted to try something out but it clashes with the bricks realistic look.

Just keeping it in here for documentation.
By the way the spike sprite is just a spike sprite from a sonic game, but I squashed it, rotated it, and cropped it.